+++ actix-http-test/CHANGES.md
@@ -1,7 +1,14 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
+
+
+## 3.0.0-beta.6 - 2021-11-15
+* `TestServer::stop` is now async and will wait for the server and system to shutdown. [#2442]
+* Update `actix-server` to `2.0.0-beta.9`. [#2442]
 * Minimum supported Rust version (MSRV) is now 1.52.
+
+[#2442]: https://github.com/actix/actix-web/pull/2442
 
 
 ## 3.0.0-beta.5 - 2021-09-09